pyproject.toml
[project]
name = "earthquake-de-project"
version = "0.1.0"
description = "Earthquake pipeline project"
authors = [{ name = "s-yu-dev", email = "s.yukhnovets@proton.me" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
    "apache-airflow==2.10.3",
    "duckdb>=1.0.0",
    "pendulum>=3.0.0",
    "s3fs",
    "psycopg2-binary>=2.9.0"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.0"
pre-commit = "^3.6.0"
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = []

[tool.ruff.lint.isort]
force-sort-within-sections = true
combine-as-imports = true
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
